Add files via upload
To execute this app just you have to add this file in the htdocs folder of the local server application folders(XAMPP , WAMP, LAMP,  or LAMP). To execute actions on the data(s), go to your phpmyadmin5 database and create a Database: E-notes with Table: _notes 
Thanks for having a look on this file!